update and fix triplification sciprts

+++ lib/sql/remove-indexes.sql
@@ -1,5 +1,5 @@
 -- nodes indexes
-drop index
+drop index if exists
 	idx_literal_lang,
 	idx_node_content,
 	idx_node_dcontent,
@@ -9,15 +9,15 @@ drop index
 -- nodes constraints
 alter table nodes drop constraint nodes_ltype_fkey;
 
-# triples indexes
-drop index
+-- triples indexes
+drop index if exists
 	-- idx_triples_spo,
 	idx_triples_cspo,
 	idx_triples_p,
 	idx_triples_ops,
-	idx_triples_pso
+	idx_triples_pso;
 
-# triples constraints
+-- triples constraints
 alter table triples drop constraint triples_context_fkey;
 alter table triples drop constraint triples_creator_fkey;
 alter table triples drop constraint triples_object_fkey;
